pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ibkr-trading-bot"
version = "1.0.0"
description = "Automated trading bot for Interactive Brokers with ML-based signal generation"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Adam K", email = "adamk@trader.cz"}
]
keywords = ["trading", "ml", "xgboost", "ibkr", "gui", "backtesting"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "numpy>=1.25.0",
    "pandas>=2.2.0",
    "pyarrow>=14.0.0",
    "PySide6>=6.7.0",
    "matplotlib>=3.8.0",
    "plotly>=5.20.0",
    "seaborn>=0.13.0",
    "scikit-learn>=1.4.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.3.0",
    "shap>=0.44.0",
    "scipy>=1.13.0",
    "joblib>=1.4.0",
    "ib-insync>=0.9.84",
    "pyyaml>=6.0.1",
    "tqdm>=4.66.0",
    "loguru>=0.7.2",
    "tvDatafeed>=1.3.7",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.15.0",
    "mypy>=1.19.0",
    "types-PyYAML",
    "pandas-stubs",
    "pre-commit>=3.5.0",
]
dl = [
    "tensorflow>=2.16.0",
    "torch>=2.2.0",
]

[tool.setuptools]
packages = ["ibkr_trading_bot"]

# ==================== Ruff Configuration ====================
[tool.ruff]
line-length = 120
target-version = "py310"
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    ".eggs",
    "*.egg-info",
    ".pytest_cache",
    ".mypy_cache",
    "xtest*.py",  # Exclude test/helper scripts
]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "RUF",  # Ruff-specific
    "UP",   # pyupgrade
]
ignore = [
    "E501",   # Line too long (delegated to formatter)
    "W505",   # Doc line too long
]

[tool.ruff.lint.isort]
known-first-party = ["ibkr_trading_bot"]
combine-as-imports = true

# ==================== Black Configuration ====================
[tool.black]
line-length = 120
target-version = ["py310", "py311", "py312", "py313"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.pytest_cache
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

# ==================== MyPy Configuration ====================
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_optional = true
exclude = [
    "xtest*.py",
    "tests/",
]

[[tool.mypy.overrides]]
module = [
    "pandas.*",
    "sklearn.*",
    "xgboost.*",
    "lightgbm.*",
    "matplotlib.*",
    "plotly.*",
    "seaborn.*",
    "joblib.*",
    "scipy.*",
    "tvDatafeed.*",
]
ignore_missing_imports = true

# ==================== Pytest Configuration ====================
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["ibkr_trading_bot"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__main__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
skip_covered = false
